issue-2725: mv actor to libs storage core
