+++ cloud/blockstore/libs/storage/core/ya.make
@@ -3,6 +3,7 @@ LIBRARY()
 GENERATE_ENUM_SERIALIZATION(mount_token.h)
 
 SRCS(
+    acquire_disk_actor.cpp
     block_handler.cpp
     compaction_map.cpp
     compaction_options.cpp
@@ -19,6 +20,7 @@ SRCS(
     pending_request.cpp
     probes.cpp
     proto_helpers.cpp
+    release_disk_actor.cpp
     request_buffer.cpp
     request_info.cpp
     storage_request_counters.cpp
