+++ cloud/blockstore/libs/storage/disk_registry/disk_registry_private.h
@@ -5,6 +5,7 @@
 #include <cloud/blockstore/libs/kikimr/components.h>
 #include <cloud/blockstore/libs/kikimr/events.h>
 #include <cloud/blockstore/libs/storage/api/disk_agent.h>
+#include <cloud/blockstore/libs/storage/core/acquire_release_disk.h>
 #include <cloud/blockstore/libs/storage/core/request_info.h>
 #include <cloud/blockstore/libs/storage/protos/disk.pb.h>
 
@@ -104,8 +105,9 @@ struct TCachedAcquireKey
     }
 };
 
-using TCachedAcquireRequests =
-    TMap<TCachedAcquireKey, TAgentAcquireDevicesCachedRequest>;
+using TCachedAcquireRequests = TMap<
+    TCachedAcquireKey,
+    NAcquireReleaseDevices::TAgentAcquireDevicesCachedRequest>;
 
 ////////////////////////////////////////////////////////////////////////////////
